BigInt

+++ GuessWho/model.py
@@ -5,9 +5,9 @@ from flask_sqlalchemy import SQLAlchemy
 db = SQLAlchemy()
 
 class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     username = db.Column(db.String(80), unique=True, nullable=False)
-    follower_count = db.Column(db.Integer, nullable=False)
+    follower_count = db.Column(db.BigInteger, nullable=False)
     #Tweet IDs are ordinal ints, so we can fetch most recent tweets
     newest_tweet_id = db.Column(db.BigInteger, nullable=False)
 
@@ -16,7 +16,7 @@ class User(db.Model):
 
         
 class Tweet(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     text = db.Column(db.Unicode(300))
     embedding =db.Column(db.PickleType, nullable=False)
     user_id = db.Column(db.BigInteger, db.ForeignKey('user.id'), nullable=False)
